refactor(atlassian): restructure setup web server module and enhance routing functionality
This commit reorganizes the setup web server module by introducing a new barrel file for better import management and creating a dedicated route context interface to prevent circular dependencies. It also adds several utility functions for handling JSON responses, parsing request bodies, and managing credentials. New route handlers for GET, POST, and status checks are implemented to streamline the setup process. Additionally, documentation is added to clarify the module's structure and purpose, improving maintainability and user guidance.
