+++ packages/atlassian/src/mcp/tools/setup/web-server/constants/mask.ts
@@ -1,0 +1,3 @@
+/** Placeholder shown in setup UI responses to indicate a credential is set
+ *  without revealing the actual value. Ten U+2022 bullets. */
+export const MASK = '••••••••••';
