+++ packages/atlassian/src/mcp/tools/setup/web-server/index.ts
@@ -1,0 +1,4 @@
+export { startSetupServer } from './web-server.js';
+export type { SetupServerOptions } from './web-server.js';
+export { createRouteHandler } from './routes.js';
+export type { RouteContext } from './route-context.js';
