+++ packages/atlassian/src/mcp/tools/setup/web-server/utils/send-json.ts
@@ -1,0 +1,10 @@
+import type { ServerResponse } from 'node:http';
+
+export function sendJson(res: ServerResponse, status: number, data: unknown): void {
+  const body = JSON.stringify(data);
+  res.writeHead(status, {
+    'Content-Type': 'application/json',
+    'Content-Length': Buffer.byteLength(body),
+  });
+  res.end(body);
+}
